Address peter's comment
